harness: extract invocation inspect callback into callback mod
