+++ harness/src/callback/mod.rs
@@ -1,0 +1,4 @@
+//! Callbacks used to hook into the invocation pipeline.
+
+#[cfg(feature = "invocation-inspect-callback")]
+pub mod invocation_inspect;
